Add support for /files/
/files/ isn't really part of the API, but it's easy enough to include
it.  Had to branch the curl setup to compare API vs files, but the use
for the end user is the same.

anything starting with /files/ requires `-o` flag for destination.
TODO: Check w/ the Files API if the /files/ path is used for creating
files as well.  That could be a problem, in which case we'll just
remove the requirement for `-o` instead of further parsing which method
is being called.